+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/inspections/GitlabLintInspection.kt
@@ -1,7 +1,7 @@
 package com.github.blarc.gitlab.template.lint.plugin.inspections
 
 import com.github.blarc.gitlab.template.lint.plugin.GitlabLintUtils
-import com.github.blarc.gitlab.template.lint.plugin.runLinting
+import com.github.blarc.gitlab.template.lint.plugin.lintGitlabYaml
 import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
 import com.github.blarc.gitlab.template.lint.plugin.settings.lintFrequency.LintFrequencyEnum
 import com.github.blarc.gitlab.template.lint.plugin.widget.LintStatusEnum
@@ -18,7 +18,7 @@ class GitlabLintInspection : LocalInspectionTool() {
         return object : PsiElementVisitor() {
             override fun visitFile(file: PsiFile) {
                 if (AppSettings.instance.lintFrequency == LintFrequencyEnum.ON_CHANGE) {
-                    runLinting(file)
+                    lintGitlabYaml(file)
                 }
                 else {
                     GitlabLintUtils.updateStatusWidget(file.project, LintStatusEnum.WAITING)