+++ src/main/kotlin/com/github/blarc/gitlab/template/lint/plugin/listeners/SaveActionListener.kt
@@ -1,6 +1,6 @@
 package com.github.blarc.gitlab.template.lint.plugin.listeners
 
-import com.github.blarc.gitlab.template.lint.plugin.runLinting
+import com.github.blarc.gitlab.template.lint.plugin.lintGitlabYaml
 import com.github.blarc.gitlab.template.lint.plugin.settings.AppSettings
 import com.github.blarc.gitlab.template.lint.plugin.settings.lintFrequency.LintFrequencyEnum
 import com.intellij.openapi.editor.Document
@@ -15,7 +15,7 @@ class SaveActionListener(val project: Project) : FileDocumentManagerListener {
         DumbService.getInstance(project).runWhenSmart {
             val file = PsiDocumentManager.getInstance(project).getPsiFile(document)
             if (file != null && AppSettings.instance.lintFrequency == LintFrequencyEnum.ON_SAVE) {
-                runLinting(file)
+                lintGitlabYaml(file)
             }
             super.beforeDocumentSaving(document)
         }